+++ Libraries/LightningShaders/LightningShaderGlslBackend.cpp
@@ -29,7 +29,7 @@ bool LightningShaderGlslBackend::RunTranslationPass(ShaderTranslationPassResult&
 
   ShaderByteStream& inputByteStream = inputData.mByteStream;
   uint32_t* data = (uint32_t*)inputByteStream.Data();
-  uint32 wordCount = inputByteStream.WordCount();
+  size_t wordCount = inputByteStream.WordCount();
 
   spirv_cross::CompilerGLSL compiler(data, wordCount);
   // Set options
@@ -261,7 +261,7 @@ void LightningShaderGlslBackend::PopulateTypeName(GlslBackendInternalData& inter
 
 void LightningShaderGlslBackend::PopulateMemberTypeInfo(GlslBackendInternalData& internalData,
                                                     spirv_cross::SPIRType& parentType,
-                                                    int memberIndex,
+                                                    u32 memberIndex,
                                                     ShaderResourceReflectionData& reflectionData,
                                                     bool isInterfaceType)
 {
@@ -270,7 +270,7 @@ void LightningShaderGlslBackend::PopulateMemberTypeInfo(GlslBackendInternalData&
   spirv_cross::SPIRType memberType = compiler->get_type(parentType.member_types[memberIndex]);
   PopulateTypeName(internalData, memberType, reflectionData);
   reflectionData.mInstanceName = compiler->get_member_name(parentType.self, memberIndex).c_str();
-  reflectionData.mSizeInBytes = compiler->get_declared_struct_member_size(parentType, memberIndex);
+  reflectionData.mSizeInBytes = (u32)compiler->get_declared_struct_member_size(parentType, memberIndex);
 
   reflectionData.mStride = 0;
   // Get array stride
@@ -303,7 +303,7 @@ void LightningShaderGlslBackend::PopulateTypeInfo(GlslBackendInternalData& inter
   // type's size and whether this is a matrix or vector
   if (isInterfaceType)
   {
-    size_t baseTypeByteSize = spirvType.width / 8;
+    u32 baseTypeByteSize = spirvType.width / 8;
     reflectionData.mSizeInBytes = spirvType.columns * spirvType.vecsize * baseTypeByteSize;
   }
   else
@@ -320,7 +320,7 @@ void LightningShaderGlslBackend::ExtractResourceReflection(GlslBackendInternalDa
   String instanceName = compiler->get_name(resource.id).c_str();
 
   // Find the reflection data by name instance name
-  int index = results.Size();
+  size_t index = results.Size();
   for (size_t i = 0; i < results.Size(); ++i)
   {
     if (results[i].mReflectionData.mInstanceName == instanceName)
@@ -355,7 +355,7 @@ void LightningShaderGlslBackend::ExtractResourceReflection(GlslBackendInternalDa
   {
     ShaderResourceReflectionData& memberReflection = stageResource.mMembers[i];
     // Populate the member reflection data (offset, size, etc...)
-    PopulateMemberTypeInfo(internalData, spirvType, i, memberReflection, isInterfaceType);
+    PopulateMemberTypeInfo(internalData, spirvType, static_cast<u32>(i), memberReflection, isInterfaceType);
     // Add a string to member index mapping
     stageResource.mLookupMap[memberReflection.mInstanceName] = i;
   }
